refactor: move throttle method to CollectHandlers
