+++ src/Handlers/CollectHandlers.php
@@ -7,6 +7,7 @@ use SergiX44\Nutgram\Exception\StatusFinalizedException;
 use SergiX44\Nutgram\Handlers\Listeners\MessageListeners;
 use SergiX44\Nutgram\Handlers\Listeners\SpecialListeners;
 use SergiX44\Nutgram\Handlers\Listeners\UpdateListeners;
+use SergiX44\Nutgram\Middleware\RateLimit;
 use SergiX44\Nutgram\Telegram\Types\Common\Update;
 
 abstract class CollectHandlers
@@ -42,6 +43,19 @@ abstract class CollectHandlers
      * @var bool
      */
     protected bool $finalized = false;
+
+    protected ?RateLimit $rateLimit = null;
+
+    public function throttle(int $maxAttempts, int $decaySeconds = 60): self
+    {
+        $this->rateLimit = new RateLimit(
+            maxAttempts: $maxAttempts,
+            decaySeconds: $decaySeconds,
+            key: 'global',
+        );
+
+        return $this;
+    }
 
     /**
      * @param callable|callable-string|array $callable
